make create_git_repo() a fixture and create subprocess_utils
